read socket storage directly in res.destroy; move tests to their own file
The ServerResponse socket getter auto-creates a FakeSocket, so the
no-socket fallback was unreachable and a standalone new ServerResponse()
never emitted 'close' after destroy(). Read this[fakeSocketSymbol]
directly (same pattern as end()/setTimeout()) and schedule emitCloseNT
whenever there is no native handle, which covers the FakeSocket case.

Also move the tests to test/js/node/http/node-http-server-response-destroy.test.ts
so the gate runs them in isolation, and add coverage for the standalone
response path.
